rename and refactor composable
