+++ src/modules/data/board/index.ts
@@ -9,7 +9,7 @@ import * as cardActions from "./cardActions/cardActions";
 import { useSocketConnection } from "./socket/socket";
 import { useCardStore } from "./Card.store";
 import { useBoardInactivity } from "./boardInactivity.composable";
-import { useForceRender } from "./forceRenderComponent.composable";
+import { useForceRender } from "./fixSamePositionDnD.composable";
 
 export {
 	boardActions,
